+++ src/pogo-utils/heading_PID.h
@@ -27,7 +27,7 @@
  *   3) heading_pid_set_limits(&pid, 1.0, 0.2);  // u_max, I_max (normalized)
  *   4) heading_pid_set_target(&pid, psi_target);
  *   5) heading_pid_enable(&pid, true);
- *   6) double u = heading_pid_update(&pid);     // call every control tick
+ *   6) float u = heading_pid_update(&pid);     // call every control tick
  *   7) Map 'u' to motors (e.g., differential steering).
  */
 
@@ -35,17 +35,17 @@
 typedef struct {
     // --- configuration ---
     const heading_detection_t *hd;  ///< Linked heading-detection config/state (may be NULL if using *from_heading* API only)
-    double Kp;                      ///< Proportional gain
-    double Ki;                      ///< Integral gain
-    double Kd;                      ///< Derivative gain
-    double u_max;                   ///< Absolute output saturation (|u| <= u_max, typically <= 1.0)
-    double I_max;                   ///< Absolute integrator clamp (|I| <= I_max)
+    float Kp;                      ///< Proportional gain
+    float Ki;                      ///< Integral gain
+    float Kd;                      ///< Derivative gain
+    float u_max;                   ///< Absolute output saturation (|u| <= u_max, typically <= 1.0)
+    float I_max;                   ///< Absolute integrator clamp (|I| <= I_max)
 
     // --- target/state ---
-    double target_rad;              ///< Target heading (radians), wrapped to (-pi, pi]
-    double e_prev;                  ///< Previous error for derivative term
-    double I;                       ///< Integrator state
-    double u;                       ///< Last computed control (post-saturation)
+    float target_rad;              ///< Target heading (radians), wrapped to (-pi, pi]
+    float e_prev;                  ///< Previous error for derivative term
+    float I;                       ///< Integrator state
+    float u;                       ///< Last computed control (post-saturation)
     uint32_t t_prev_ms;             ///< Timestamp of last update (ms)
     bool enabled;                   ///< If false, update() returns 0 and leaves motors alone
     bool first_update;              ///< To skip D term spike on first call
@@ -66,21 +66,21 @@ void heading_pid_enable(heading_pid_t *pid, bool enable);
 void heading_pid_reset(heading_pid_t *pid);
 
 /** @brief Set PID gains. */
-void heading_pid_set_gains(heading_pid_t *pid, double Kp, double Ki, double Kd);
+void heading_pid_set_gains(heading_pid_t *pid, float Kp, float Ki, float Kd);
 
 /** @brief Set output and integrator limits (|u|<=u_max, |I|<=I_max). */
-void heading_pid_set_limits(heading_pid_t *pid, double u_max, double I_max);
+void heading_pid_set_limits(heading_pid_t *pid, float u_max, float I_max);
 
 /** @brief Set/overwrite the target heading (radians). */
-void heading_pid_set_target(heading_pid_t *pid, double target_rad);
+void heading_pid_set_target(heading_pid_t *pid, float target_rad);
 
 /** @brief Get latest normalized control output in [-u_max,+u_max]. */
-static inline double heading_pid_get_output(const heading_pid_t *pid) {
+static inline float heading_pid_get_output(const heading_pid_t *pid) {
     return pid ? pid->u : 0.0;
 }
 
 /** @brief Get latest error (target−measured) after wrapping to (-pi, pi]. */
-static inline double heading_pid_get_error(const heading_pid_t *pid) {
+static inline float heading_pid_get_error(const heading_pid_t *pid) {
     return pid ? pid->e_prev : 0.0;
 }
 
@@ -92,7 +92,7 @@ static inline double heading_pid_get_error(const heading_pid_t *pid) {
  *
  * @return Saturated control 'u' in [-u_max, +u_max]. Returns 0 if disabled or on null ptrs.
  */
-double heading_pid_update(heading_pid_t *pid);
+float heading_pid_update(heading_pid_t *pid);
 
 /**
  * @brief Update from a provided measured heading (radians).
@@ -101,7 +101,7 @@ double heading_pid_update(heading_pid_t *pid);
  *
  * @return Saturated control 'u' in [-u_max, +u_max]. Returns 0 if disabled or on null ptrs.
  */
-double heading_pid_update_from_heading(heading_pid_t *pid, double measured_heading_rad);
+float heading_pid_update_from_heading(heading_pid_t *pid, float measured_heading_rad);
 
 
 #endif /* HEADING_PID_H */
